Rename command to be more descriptive
